requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv>=1.0.0
# NLP
spacy>=3.7.0
nltk>=3.8.0
# File extraction
pymupdf>=1.24.0
python-docx>=1.1.0
# LLM (legal reasoning - optional)
openai>=1.0.0
anthropic>=0.18.0
# PDF export
reportlab>=4.0.0
